+++ SS14.MapServer/Services/Github/GithubApiService.cs
@@ -32,7 +32,7 @@ public sealed class GithubApiService : AbstractGithubApiService
         return compareResult.Files.Select(file => file.Filename);
     }
 
-    public async Task<int?> CreateCommentWithTemplate(InstallationIdentifier installation, IssueIdentifier issue, string templateName, object model)
+    public async Task<long?> CreateCommentWithTemplate(InstallationIdentifier installation, IssueIdentifier issue, string templateName, object model)
     {
         if (!await CheckRateLimit(installation))
             return null;
@@ -51,7 +51,7 @@ public sealed class GithubApiService : AbstractGithubApiService
         return null;
     }
 
-    public async Task UpdateCommentWithTemplate(InstallationIdentifier installation, IssueIdentifier issue, int commentId, string templateName, object model)
+    public async Task UpdateCommentWithTemplate(InstallationIdentifier installation, IssueIdentifier issue, long commentId, string templateName, object model)
     {
         if (!await CheckRateLimit(installation))
             return;
